kei: split into lib + bin; add __fuzz_internals feature
Move every module declaration, helper, and the run() function out of
main.rs into a new src/lib.rs. main.rs is now a 7-line shim that calls
kei::main_inner(). Behavior is unchanged - this is purely a structural
move so the module tree is reachable as a library, not just through the
binary.

Why now: the fuzz harnesses in fuzz/ have been inlining their target
modules via #[path] because a binary-only crate has no library target
for fuzz/Cargo.toml to depend on. That works for leaf modules but
breaks for any parser whose module imports `crate::xxx` or
`super::xxx` - which locks out config.rs (TOML), auth/responses.rs
(SRP), icloud/photos/asset.rs (PhotoAsset construction), and
state/types.rs (FromStr enums) from getting fuzz coverage.

Also adds an __fuzz_internals Cargo feature that gates a
`pub mod __fuzz` in lib.rs. The module exposes pub fn wrappers around
the parser entry points - taking and returning only externally
nameable types (bytes, strings, serde_json::Value), discarding typed
results internally. Internal types stay pub(crate); this is the one
place that intentionally widens visibility, and only when the feature
is on.

Mechanical changes in main.rs: zero crate:: references existed before
the move (all internal sibling-module access went through unprefixed
paths like `password::SecretString`), so the lib.rs body is byte-for-byte
identical to the old main.rs except for the rename of `fn main` to
`pub fn main_inner`.

justfile: switch test recipes from `cargo test --bin kei` to
`cargo test --lib`, and the live-coverage step likewise. The binary
now contains one function call so there's nothing to test in --bin.

Verified: just gate passes (fmt + clippy + lib + cli + behavioral +
doc + audit + typos). No tests changed locations or behavior.
